♻️ [Refactor] 홈 뷰(캘린더, 미리보기) 리팩토링
