+++ Diary/Diary/Models/ReflectionStatus.swift
@@ -1,0 +1,12 @@
+//
+//  ReflectionStatus.swift
+//  Diary
+//
+//  Created by Yurim on 7/25/25.
+//
+
+enum ReflectionStatus {
+  case none      // 일기 없음
+  case pending   // 회고 없음 ("-")
+  case completed // 회고 완료
+}
