+++ Diary/Diary/Presentations/Diary/DiaryDetail/Views/DiaryDetailView.swift
@@ -5,14 +5,13 @@
 //  Created by jinhyeokKim on 7/17/25.
 //
 
-import SwiftUI
+import Foundation
 
-struct DiaryDetailView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    DiaryDetailView()
+struct DiaryEntry: Identifiable {
+  let id = UUID()
+  let date: Date
+  let content: String
+  let quote: String
+  let vow: String
+  let reflection: String
 }
