--- Diary/Diary/Presentations/Home/ViewModels/DiaryEntry.swift
@@ -1,17 +1,0 @@
-//
-//  DiaryEntry.swift
-//  Diary
-//
-//  Created by Yurim on 7/24/25.
-//
-
-import Foundation
-
-struct DiaryEntry: Identifiable {
-  let id = UUID()
-  let date: Date
-  let content: String
-  let quote: String
-  let vow: String
-  let reflection: String
-}
